+++ src/brod_consumer.erl
@@ -22,6 +22,7 @@
         , start_link/4
         , start_link/5
         , stop/1
+        , stop_maybe_kill/2
         , subscribe/3
         , unsubscribe/2
         ]).
@@ -67,7 +68,7 @@
 
 -type pending_acks() :: #pending_acks{}.
 
--record(state, { client_pid          :: pid()
+-record(state, { bootstrap           :: pid() | brod:bootstrap()
                , connection          :: ?undef | pid()
                , topic               :: binary()
                , partition           :: integer()
@@ -112,10 +113,10 @@
 
 %%%_* APIs =====================================================================
 %% @equiv start_link(ClientPid, Topic, Partition, Config, [])
--spec start_link(pid(), topic(), partition(), config()) ->
+-spec start_link(pid() | brod:bootstrap(), topic(), partition(), config()) ->
         {ok, pid()} | {error, any()}.
-start_link(ClientPid, Topic, Partition, Config) ->
-  start_link(ClientPid, Topic, Partition, Config, []).
+start_link(Bootstrap, Topic, Partition, Config) ->
+  start_link(Bootstrap, Topic, Partition, Config, []).
 
 %% @doc Start (link) a partition consumer.
 %% Possible configs:
@@ -157,14 +158,27 @@ start_link(ClientPid, Topic, Partition, Config) ->
 %%     then let it dynamically adjust to a number around
 %%     PrefetchCount * AverageSize
 %% @end
--spec start_link(pid(), topic(), partition(), config(), [any()]) ->
+-spec start_link(pid() | brod:bootstrap(),
+                 topic(), partition(), config(), [any()]) ->
                     {ok, pid()} | {error, any()}.
-start_link(ClientPid, Topic, Partition, Config, Debug) ->
-  Args = {ClientPid, Topic, Partition, Config},
+start_link(Bootstrap, Topic, Partition, Config, Debug) ->
+  Args = {Bootstrap, Topic, Partition, Config},
   gen_server:start_link(?MODULE, Args, [{debug, Debug}]).
 
 -spec stop(pid()) -> ok | {error, any()}.
 stop(Pid) -> safe_gen_call(Pid, stop, infinity).
+
+-spec stop_maybe_kill(pid(), timeout()) -> ok.
+stop_maybe_kill(Pid, Timeout) ->
+  try
+    gen_server:call(Pid, stop, Timeout)
+  catch
+    exit : {noproc, _} ->
+      ok;
+    exit : {timeout, _} ->
+      exit(Pid, kill),
+      ok
+  end.
 
 %% @doc Subscribe or resubscribe on messages from a partition.
 %% Caller may pass in a set of options which is an extention of consumer config
@@ -206,7 +220,8 @@ debug(Pid, File) when is_list(File) ->
 
 %%%_* gen_server callbacks =====================================================
 
-init({ClientPid, Topic, Partition, Config}) ->
+init({Bootstrap, Topic, Partition, Config}) ->
+  process_flag(trap_exit, true),
   Cfg = fun(Name, Default) ->
           proplists:get_value(Name, Config, Default)
         end,
@@ -218,8 +233,14 @@ init({ClientPid, Topic, Partition, Config}) ->
   PrefetchBytes = erlang:max(Cfg(prefetch_bytes, ?DEFAULT_PREFETCH_BYTES), 0),
   BeginOffset = Cfg(begin_offset, ?DEFAULT_BEGIN_OFFSET),
   OffsetResetPolicy = Cfg(offset_reset_policy, ?DEFAULT_OFFSET_RESET_POLICY),
-  ok = brod_client:register_consumer(ClientPid, Topic, Partition),
-  {ok, #state{ client_pid          = ClientPid
+  %% If bootstrap is a client pid, register self to the client
+  case is_shared_conn(Bootstrap) of
+    true ->
+      ok = brod_client:register_consumer(Bootstrap, Topic, Partition);
+    false ->
+      ok
+  end,
+  {ok, #state{ bootstrap           = Bootstrap
              , topic               = Topic
              , partition           = Partition
              , begin_offset        = BeginOffset
@@ -266,10 +287,12 @@ handle_info({'DOWN', _MonitorRef, process, Pid, _Reason},
                                      }),
   {noreply, NewState};
 handle_info({'DOWN', _MonitorRef, process, Pid, _Reason},
-            #state{connection = Pid} = State0) ->
-  State = State0#state{connection = ?undef, connection_mref = ?undef},
-  ok = maybe_send_init_connection(State),
-  {noreply, State};
+            #state{connection = Pid} = State) ->
+  %% monitored connection managed by brod_client
+  {noreply, handle_conn_down(State)};
+handle_info({'EXIT', Pid, _Reason}, #state{connection = Pid} = State) ->
+  %% standalone connection spawn-linked to self()
+  {noreply, handle_conn_down(State)};
 handle_info(Info, State) ->
   error_logger:warning_msg("~p ~p got unexpected info: ~p",
                           [?MODULE, self(), Info]),
@@ -317,15 +340,31 @@ handle_cast(Cast, State) ->
                           [?MODULE, self(), Cast]),
   {noreply, State}.
 
-terminate(Reason, _State) ->
-  error_logger:warning_msg("~p ~p terminating, reason:\n~p",
-                           [?MODULE, self(), Reason]),
-  ok.
+terminate(Reason, #state{topic = T,
+                         partition = P,
+                         connection = C
+                        } = State) ->
+  case is_shared_conn(State) of
+    true -> ok;
+    false -> kpro:close_connection(C)
+  end,
+  is_normal(Reason) orelse
+  error_logger:info_msg("Consumer ~s-~w terminate reason: ~p", [T, P, Reason]).
 
 code_change(_OldVsn, State, _Extra) ->
   {ok, State}.
 
 %%%_* Internal Functions =======================================================
+
+is_normal(normal) -> true;
+is_normal(shutdown) -> true;
+is_normal({shutdown, _}) -> true;
+is_normal(_) -> false.
+
+handle_conn_down(State0) ->
+  State = State0#state{connection = ?undef, connection_mref = ?undef},
+  ok = maybe_send_init_connection(State),
+  State.
 
 do_debug(Pid, Debug) ->
   {ok, _} = gen:call(Pid, system, {debug, Debug}, infinity),
@@ -718,15 +757,19 @@ safe_gen_call(Server, Call, Timeout) ->
 %% Init payload connection regardless of subscriber state.
 -spec maybe_init_connection(state()) ->
         {ok, state()} | {{error, any()}, state()}.
-maybe_init_connection(#state{ client_pid = ClientPid
-                        , topic      = Topic
-                        , partition  = Partition
-                        , connection = ?undef
-                        } = State0) ->
+maybe_init_connection(
+  #state{ bootstrap  = Bootstrap
+        , topic      = Topic
+        , partition  = Partition
+        , connection = ?undef
+        } = State0) ->
   %% Lookup, or maybe (re-)establish a connection to partition leader
-  case brod_client:get_leader_connection(ClientPid, Topic, Partition) of
+  case connect_leader(Bootstrap, Topic, Partition) of
     {ok, Connection} ->
-      Mref = erlang:monitor(process, Connection),
+      Mref = case is_shared_conn(Bootstrap) of
+               true -> erlang:monitor(process, Connection);
+               false -> ?undef %% linked
+             end,
       %% Switching to a new connection
       %% the response for last_req_ref will be lost forever
       State = State0#state{ last_req_ref = ?undef
@@ -741,6 +784,14 @@ maybe_init_connection(State) ->
   {ok, State}.
 
 
+connect_leader(ClientPid, Topic, Partition) when is_pid(ClientPid) ->
+  brod_client:get_leader_connection(ClientPid, Topic, Partition);
+connect_leader(Endpoints, Topic, Partition) when is_list(Endpoints) ->
+  connect_leader({Endpoints, []}, Topic, Partition);
+connect_leader({Endpoints, ConnCfg}, Topic, Partition) ->
+  %% connection pid is linked to self()
+  kpro:connect_partition_leader(Endpoints, ConnCfg, Topic, Partition).
+
 %% Send a ?INIT_CONNECTION delayed loopback message to re-init.
 -spec maybe_send_init_connection(state()) -> ok.
 maybe_send_init_connection(#state{subscriber = Subscriber}) ->
@@ -749,6 +800,10 @@ maybe_send_init_connection(#state{subscriber = Subscriber}) ->
   brod_utils:is_pid_alive(Subscriber) andalso
     erlang:send_after(Timeout, self(), ?INIT_CONNECTION),
   ok.
+
+%% In case 'bootstrap' is a client pid, connection is shared with other workers.
+is_shared_conn(#state{bootstrap = Bootstrap}) -> is_shared_conn(Bootstrap);
+is_shared_conn(Bootstrap) -> is_pid(Bootstrap).
 
 %%%_* Tests ====================================================================